code.txt
Super Market cash Counter program. 

Program starts with declaring and initializing values to n, w, n1 and declaring a array 'u' of of size 100 and a string s having limit n containing names of the components to be bought by the customer. 
Then the customer is asked to enter the quantity of the product he/she has bought. Then price of product will be entered by the cashier. 

After that the customer is asked whether he/she wants to delete, add, replace, change the quantity of the products. 
Else the bill will be created for the customer or the cashier can end the program by pressing no.  7 and a "thank you" message will be printed.